Adds the forcing class to the hybrid model
The forcing for hybrid RANS/LES is fully implemented in this commit. But
it still needs to be hooked up to the hybrid mediator in order to work.
Testing was also added to verify some of the low-level functions as well
as perform a smoke-test.

(cherry picked from commit 5c7f3113393f3de03d4e7b71b43ae3038274e105)