Refactor the COutput class to simplify output
Previously, coders had to add code to 7 different locations inside
`COutput` in order to add a new variable to the output.  This made for
a lot of code duplication and wasted time when adding a new variable.
This commit utilizies function pointers to reduce variable registration
to a single line.

This is a work in progress, and there are three changes that will be
addressed in future commits:

1. Registration for multiple zones
2. Generalization to tensor variables
3. Generalization to geometry variables (CPoint instead of CNode)