Fix bug in v2f restart
After restarting a v2-f caculations, residuals would jump and
the calculation would not resume in the same place.  This is
because the eddy viscosity is computed in the postprocessing
script, but the turbulent timescale is calculated elsewhere.
This meant that, upon restart, the turbulent timescale had
its default value.

This commit forces the turbulent timescale calculation to
update during postprocessing.