Move RK defaults from `SetOutput` to `SetPostproc`
This fixes a bug in the RK LIMEX EDIRK setup. Previously, if no
coefficients were specified a sensible set of defaults were specified
for explicit RK and RK LIMEX EDIRK.  These were set in `SetOutput`,
which was not a suitable place for setup. `SetOutput` is only run on the
master task.  This caused explicit RK and RK LIMEX EDIRK to fail in
parallel, due to undefined RK variables.

This change moves the RK default initialization to `SetPostprocessing`,
which is run by all MPI tasks.