Add damping of SGET term in high AR cells.
This commit is a work-in-progress, though it compiles and runs fine.
The problem is that it is overzealous.  This fix is intended only for
BL, but the damping will happen in any high aspect-ratio cells, even in
the freestream or far away from walls.

Ideally, this term would check if the cell is "near" a wall by comparing
the dissipation or the wall-distance and only apply the damping near
walls.  But how to do that is not immediately apparent.  The value used
in SU2 for "freestream dissipation" when initializing a flow with no
restart is not a representative approximation of freestream dissipation in
a final steady-state flow.