Fix bug in FD/NTS blending for SLAU and SLAU2
Previously, the numerics for SLAU and SLAU2 expected a valid value of
the "roe low dissipation" i.e., the blending function.  This was
supposed to be set in CEulerSolver::Preprocessing or
CNSSolver::Preprocessing, but was only set for ROE (not SLAU or SLAU2).
This commit fixes this bug by adding SLAU and SLAU2 to the conditional
check.