Remove unnecessary CDP hacks and bugs
The previous commit added some wrong behavior and bugs, in order to best
imitate the bad behavior and bugs present in r8839 of CDP.  This commit
removes those bugs but leaves in place the necessary, correct changes
that make SU2 better match CDP.

The resolution tensor no longer matches that calculated by CDP,
predicting a lower value for most of the log layer.  This is primarily
due to the change from a symmetric eigenvalue calculation to a
generalized eigenvalue calculation.  The other differences had much more
minor impact.