Forcing hacked to match CDP.
This is a set of changes that bring CDP and SU2 into agreement w.r.t the
forcing.  There's a lot of messy code and hacks, so this is not intended
to be a final state.  Rather, it represents a starting point from which
intelligent choices can be made about code-cleanup and improvements.
Those improvements may cause some disagreement between SU2 and CDP, so
this commit represents the best possible agreement, which is different
from the the best possible performance or quality.