Remove extra 2/3 term from viscous numerics
This change was previously applied to the turbulent stresses used in the
resolution adequacy calculation, but was not updated here.  It is my
intention to add a later commit consolidating these calculations to
avoid repitition and protect against bugs like this.

The fluctuating stress model was formulated prior to the model split idea
coming along, and the 2/3 term was  explicitly a model for the deviatoric
part of the stress.  We wanted to ensure there would be no contribution
from the model to the isotropic part, which was explicitly represented
using the fluctuating subgrid k from the turbulence model.  This
justification doesn't exactly apply anymore because of the model split
form and because we only get mean k from the turbulence model.

In the current case, the isotropic part of the tau^{SGET} model amounts
to a model for the fluctuations in the subgrid k (assuming it is mean 0,
like in M43).  In the currently implemented form, it is zero, which is
obviously not true. But it can be used as a starting point.