Adjust adaptive CFL timestepping for v2-f
It has been observed that the f and dissipation residuals are the
hardest to converge.  This means that the density, momentum, or
kinetic energy residuals may be descreasing while the dissipation
residual is increasing.  This change makes the CFL adaptation dependent
on the dissipation residual, rather than the density residual.  This
greatly decreases the likelihood of divergence due to the turbulence
variables diverging.