Add limits to improve convergence while transient
This is a set of two related changes to the numerics, in order to
improve stability.  The first limits the value of C_e1 to values that
are realistic (and not near-infinite).  The second ensures that the
"min" limit on the production of v2 is applied correctly.  If some terms
are negative, then "min" has the opposite of the intended effect.

Both of these changes have been observed to not harm convergence for an
incompressible flat-plate, and they improve convergence for a M=2 flat
plate test-case.