Fix bug in RK IMEX schemes.
Previously, if no parameters were specified, the default RK
parameters were set in `Config::SetOutput`.  But `Config::SetOutput` is
only called by the rank 0 MPI task.  So the rest of the nodes were left
with uninitialized RK parameters.

This commit moves the default initialization of the RK schemes to a more
appropriate place.