Limit hybrid average updates to model-split cases
This commit adds two changes to increase robustness of the code.  First,
an assert is added to ensure that the hybrid mediator exists before it
is called (when the code is executed from a debug version).  Second, the
averaging of hybrid variables only occurs when a model-split
hybridization is being used.