Remove the second set of SetRoe_Dissipation
In order to conserve variables, the same numerics must be applied in the
same way on both sides of a face/edge.  This means the "Roe
dissipation," or the hybrid central/upwind blending constant, must used
in the same way on both sides of an edge.

As the solver loops through the grid, it would previously pull the "Roe
Dissipation," compute a new (face) value, and then overwrite the old
value with the new (face) value in both nodes.  Since each cell has
multiple neighboring nodes, this resulted in the "Roe Dissipation" in
each cell being changed multiple times.  With this change, there was no
guarantee that the Roe Dissipation values used on one side of a face
would be the same as the Roe Dissipation values used on the other side
of the face.

The end result was that the system did *not* conserve mass.  Turning off
hybrid upwind/central blending fixed the mass conservation problem.

By removing the overwriting step, the original cell-specific values are
stored in each node. The only downside is that face-specific values of
the Roe dissipation must be computed manually as a post-processing step.
The solver does not record the values used at each face.