Add new periodic capabilities based on su2 file
For periodic BCs, SU2 defines a set of "halo" nodes at each end.  One
set is a true halo, and matches some interior nodes.  The other set is a
group of removed nodes that match the opposite side of the domain.
While these are treated like halo nodes (they send and receive), they
are not truly halo nodes.

For doubly-periodic boundaries, such as the 3D plane channel, the
assumptions under which SU2 was formulated break down.  For the corners,
there is now one removed periodic halo edge, and three true halo edges.
Unlike a 2D channel, there is no longer a 1-1 corredpondance of removed
periodic nodes and halo nodes.

This introduces a new option to tag a periodic transform in the cfg file
with "HALO=0" and "HALO=1".  If "HALO=1", then the periodic boundary is
not a true halo. This allows the periodic transformations to be manually
specified, instead of following hard-coded assumptions.

If these tags are not present, there (should) be no change in behavior.

A future PR will make the notation a little clearer.