Initialize all new variables
These variables were previously uninitialized on the halo (periodic)
nodes. Even though these variables aren't used during the solve, they are
passed around during the output process (for restarts and solution
visualization).  This could be addressed by:

1. Fixing output so that it doesn't use halo nodes. This isn't really
what we want to do.
2. Zero-initializing the variables.  This is a poor programming practice,
since it makes a true zero indistinguishable from a zero-initialized
variables when it is later used.  Putting dummy data in a variable to
initialize it only hides problems from a static analysis tool.
3. Properly initializing the variables.

Number 3 is the long term goal.  Number 2 is what is done in this
commit.