Fix inconsistency in production.
After the averages are updated, the value of production may have been
changed.  Previously, this value was not ever pushed back to the
turbulence solution nodes (CTurbKEVariable).  This meant that the values
used for output were lagged by a timestep.

By adding an extra overwrite to the turbulence postprocessing, the new
production values are pushed to the turbulence nodes after averaging
updates.