Update test values
In k-based turbulence models, eddy viscosity depends on laminar
viscosity, which depends on temperature, which depends on turbulent
kinetic energy.  By changing the order in which these are computed at
the end of each iteration, the transient turbulent variables change.

The change in test values was very small, and this is not believed to be
a regression.