Setup SU2 to exit and write when state is invalid
For steady-state and dual-time-stepping, it is assumed
that any invalid states are transients and can be safely ignored.
But for time-stepping (and maybe some cases of dual-time-stepping), an
invalid state indicates that the simulations has gone bad.  In those
cases, it is best to just error out.

By "invalid state", I mean cases like negative density or negative
temperature.