Tolerate errors for steady and DT routines.
For dual-time-stepping and steady routines, we can allow the unphysical
points to be replaced by the previous solution.  In these routines,
it can be assumed that these points will be replaced with a better
solution in a later (inner) iteration.  The only time we really need to
worry about unphysical points is at the outer timestep.

This change allows the old, upstream behavior for steady and DT
routines, while keeping the "don't hide unphysical points" behavior for
time-stepping.