Change Euler wall and Sym plane for non-flat walls
BC_Sym_Plane was written with only straight lines or planes as symmetry
boundaries in mind. Therefore a constant unit normal was used. If
symmetry as a slip wall in viscous flows is used the boundary in
MARKER_SYM can have some curvature which results in a non-constant unit
normal. This is fixed in this PR.

+ MARKER_SYM now allows non-straight walls with local symmetry enforced
+ For viscous flows, MARKER_EULER now accounts for viscosity
+ MARKER_EULER and MARKER_SYM (symmetry planes and Euler walls) now
  share their implementation

Right now, Euler & Sym can be used interchangeably in both (incomp. &
comp.) solvers. One could safely remove one but keeping both is
reasonable as both BC are expected by users and maybe the implementation
deviates in the future.  Both will give the same results, independent of
the 'curvyness' of the boundary (as that is checked in a preprocessing
step).

See the following issues/PRs for more details:

https://github.com/su2code/SU2/issues/735
https://github.com/su2code/SU2/pull/740