Add a user-tunable parameter to the Rf mod
As found in the write-up, Sigfried's Rf modification to the f equation
affects converged solutions to fully-developed channel flow.  It is too
restrictive across the boundary layer.  However, this limit can be
relaxed by exchanging the model coefficient "3" to a lower number, such
as 1.5 to 2. These values do not affect converged, steady profiles.