Fix bug in user-specified inlets for v2-f model
Previously, the f residual was forcing the inlet f to the freestream
value, instead of the inlet value.  This fixes that.

It is necessary to treat f differently than k, v2, or epsilon because
f is governed by an elliptic PDE, and relies on diffusion.  But
diffusion is not used at inlets, for stability reasons.